LPS-87465 Move production property to the executeNPM task
